reorganization of components
