+++ src/client/App.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { createRoot } from 'react-dom/client';
 import localStorageAvailable from '../util/localStorageAvail';
-import Board from './components/board';
+import Board from './components/ game_elements/board';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider, createTheme, styled } from '@mui/material/styles';
 import { useTheme, Switch } from '@mui/material';
