+++ src/client/components/ game_elements/row.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { RowProps } from '../../types';
+import { RowProps } from '../../../types';
 import Box from './box';
-import phrases from '../../lib/phrases';
+import phrases from '../../../lib/phrases';
 import { Stack } from '@mui/material';
 
 function Row(props: RowProps) {
